dashboard Agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
requests==2.32.5
joblib==1.3.2
python-multipart==0.0.6
pydantic==2.5.0
